backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.5
jinja2==3.1.2
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic~=2.11.7
httpx==0.25.2
langchain-core
langgraph~=0.5.3
requests~=2.32.4
starlette~=0.27.0
beautifulsoup4~=4.13.4
google-generativeai==0.8.5
pymongo==4.6.0
anthropic>=0.34.0
groq>=0.5.0
GitPython>=3.1.40
PyGithub>=2.1.1